requirements.txt
# Core Framework
python-dotenv>=1.0.0
psutil>=5.9.0
requests>=2.31.0
urllib3>=2.0.0
tenacity>=8.2.0              # Smart retry logic

# Web Scraping & Browser Automation
selenium>=4.0.0
webdriver-manager>=4.0.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
playwright>=1.40.0           # Modern browser automation

# Data Processing & Science
pandas>=2.0.0
numpy>=1.24.0
openpyxl>=3.1.0              # Excel support
python-dateutil>=2.8.2       # Date parsing
dateparser>=1.2.0            # Advanced date parsing
rapidfuzz>=3.5.0             # Fuzzy string matching
scikit-learn>=1.3.0          # Anomaly detection (IsolationForest)
pandera>=0.17.0              # Data validation schemas
datacompy>=0.11.0            # Data comparison

# Database
psycopg2-binary>=2.9.0

# PDF & Crypto
pdfplumber>=0.10.0
PyPDF2>=3.0.0
pycryptodome>=3.19.0

# UI & Progress
rich>=13.7.0                 # Terminal UI
joblib>=1.3.0                # Parallel processing helpers
tqdm>=4.66.0                 # Progress bars

# Monitoring & AI
prometheus-client>=0.19.0    # Metrics
openai>=1.0.0                # LLM integration (optional)
deep-translator>=1.11.4      # Translation services
plyer>=2.1.0                 # Desktop notifications
jinja2>=3.1.0                # Reporting templates
diskcache>=5.6.0             # File-based caching
